requirements.txt
espnet==202207
numpy==1.22.4
six
